backend/requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
hypercorn==0.15.0
quart==0.19.3
opensearch-py==2.3.1